systems/lcm: Use model value for publisher input
