+++ systems/lcm/test/lcm_subscriber_system_test.cc
@@ -8,8 +8,8 @@
 #include "drake/lcm/drake_mock_lcm.h"
 #include "drake/lcm/lcmt_drake_signal_utils.h"
 #include "drake/lcmt_drake_signal.hpp"
-#include "drake/systems/framework/test_utilities/my_vector.h"
 #include "drake/systems/lcm/lcmt_drake_signal_translator.h"
+#include "drake/systems/lcm/test/custom_drake_signal_translator.h"
 
 using drake::lcm::CompareLcmtDrakeSignalMessages;
 
@@ -208,81 +208,19 @@ GTEST_TEST(LcmSubscriberSystemTest, WaitTest) {
       future_message.get(), sample_data.value));
 }
 
-// A lcmt_drake_signal translator that preserves coordinate names.
-class CustomDrakeSignalTranslator : public LcmAndVectorBaseTranslator {
- public:
-  // An output vector type that tests that subscribers permit non-default
-  // types.
-  using CustomVector = MyVector<kDim, double>;
-
-  CustomDrakeSignalTranslator() : LcmAndVectorBaseTranslator(kDim) {}
-
-  std::unique_ptr<BasicVector<double>> AllocateOutputVector() const override {
-    return std::make_unique<CustomVector>();
-  }
-
-  static std::string NameFor(int index) {
-    return std::to_string(index) + "_name";
-  }
-
-  void Deserialize(const void* lcm_message_bytes, int lcm_message_length,
-                   VectorBase<double>* vector_base) const override {
-    CustomVector* const custom_vector =
-        dynamic_cast<CustomVector*>(vector_base);
-    ASSERT_NE(nullptr, custom_vector);
-
-    // Decode the LCM message.
-    drake::lcmt_drake_signal message;
-    int status = message.decode(lcm_message_bytes, 0, lcm_message_length);
-    ASSERT_GE(status, 0);
-    ASSERT_EQ(kDim, message.dim);
-
-    // Copy message into our custom_vector.
-    for (int i = 0; i < kDim; ++i) {
-      EXPECT_EQ(message.coord[i], NameFor(i));
-      custom_vector->SetAtIndex(i, message.val[i]);
-    }
-  }
-
-  void Serialize(double time, const VectorBase<double>& vector_base,
-                 std::vector<uint8_t>* lcm_message_bytes) const override {
-    const CustomVector* const custom_vector =
-        dynamic_cast<const CustomVector*>(&vector_base);
-    ASSERT_NE(nullptr, custom_vector);
-    ASSERT_NE(nullptr, lcm_message_bytes);
-
-    // Copy custom_vector into the message.
-    drake::lcmt_drake_signal message;
-    message.dim = kDim;
-    message.val.resize(kDim);
-    message.coord.resize(kDim);
-    message.timestamp = static_cast<int64_t>(time * 1000);
-
-    for (int i = 0; i < kDim; ++i) {
-      message.val.at(i) = custom_vector->GetAtIndex(i);
-      message.coord.at(i) = NameFor(i);
-    }
-
-    // Encode the LCM message.
-    const int lcm_message_length = message.getEncodedSize();
-    lcm_message_bytes->resize(lcm_message_length);
-    message.encode(lcm_message_bytes->data(), 0, lcm_message_length);
-  }
-};
-
 // Subscribe and output a custom VectorBase type.
 GTEST_TEST(LcmSubscriberSystemTest, CustomVectorBaseTest) {
   const std::string kChannelName = "dummy";
 
   // The "device under test" and its prerequisites.
-  CustomDrakeSignalTranslator translator;
+  test::CustomDrakeSignalTranslator translator;
   drake::lcm::DrakeMockLcm lcm;
   LcmSubscriberSystem dut(kChannelName, translator, &lcm);
 
   // Create a data-filled vector.
-  typedef CustomDrakeSignalTranslator::CustomVector CustomVector;
+  using CustomVector = test::CustomDrakeSignalTranslator::CustomVector;
   CustomVector sample_vector;
-  for (int i = 0; i < kDim; ++i) {
+  for (int i = 0; i < sample_vector.size(); ++i) {
     sample_vector.SetAtIndex(i, i);
   }
 
@@ -304,7 +242,7 @@ GTEST_TEST(LcmSubscriberSystemTest, CustomVectorBaseTest) {
   const CustomVector* const custom_output =
       dynamic_cast<const CustomVector*>(output_vector);
   ASSERT_NE(nullptr, custom_output);
-  for (int i = 0; i < kDim; ++i) {
+  for (int i = 0; i < sample_vector.size(); ++i) {
     EXPECT_EQ(sample_vector.GetAtIndex(i), custom_output->GetAtIndex(i));
   }
 }
